# src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Dictionary Backend API",
    "version": "1.0.0",
    "description": "API for dictionary, auth, word lists, and quiz word sets."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local"
    },
    {
      "url": "https://dictiionery-backend.vercel.app",
      "description": "Production"
    }
  ],
  "tags": [
    {
      "name": "Health"
    },
    {
      "name": "Dictionary"
    },
    {
      "name": "Auth"
    },
    {
      "name": "Words"
    },
    {
      "name": "Exam"
    },
    {
      "name": "Grade"
    },
    {
      "name": "Subject"
    },
    {
      "name": "Admin Uploads"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "tags": ["Health"],
        "summary": "Root health endpoint",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/hello": {
      "get": {
        "tags": ["Health"],
        "summary": "Simple hello endpoint",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/wordoftheday": {
      "get": {
        "tags": ["Dictionary"],
        "summary": "Get word of the day",
        "responses": {
          "200": {
            "description": "Word of the day",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WordOfDayResponse"
                }
              }
            }
          }
        }
      }
    },
    "/define/{word}": {
      "get": {
        "tags": ["Dictionary"],
        "summary": "Get details for a word",
        "parameters": [
          {
            "name": "word",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Word found"
          },
          "404": {
            "description": "Word not found"
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registered"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logged in"
          }
        }
      }
    },
    "/auth/google": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login via Google ID token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["idToken"],
                "properties": {
                  "idToken": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logged in"
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Get current user",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user"
          }
        }
      }
    },
    "/words": {
      "get": {
        "tags": ["Words"],
        "summary": "List words with pagination",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Words list"
          }
        }
      },
      "delete": {
        "tags": ["Words"],
        "summary": "Delete a word (admin)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "word",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deletion result"
          }
        }
      }
    },
    "/exam": {
      "get": {
        "tags": ["Exam"],
        "summary": "Get exam words",
        "parameters": [
          {
            "name": "exam",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Exam words"
          }
        }
      }
    },
    "/grade": {
      "get": {
        "tags": ["Grade"],
        "summary": "Get grade words",
        "parameters": [
          {
            "name": "grade",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Grade words"
          }
        }
      }
    },
    "/subject": {
      "get": {
        "tags": ["Subject"],
        "summary": "Get subject words",
        "parameters": [
          {
            "name": "subject",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Subject words"
          }
        }
      }
    },
    "/api/upload-excel": {
      "post": {
        "tags": ["Admin Uploads"],
        "summary": "Bulk upload words from file (admin)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file"],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "promptStyle": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Upload processed"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "RegisterRequest": {
        "type": "object",
        "required": ["username", "email", "password"],
        "properties": {
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["identifier", "password"],
        "properties": {
          "identifier": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "WordOfDayResponse": {
        "type": "object",
        "properties": {
          "word": {
            "type": "string"
          },
          "meaning": {
            "type": "string"
          },
          "date": {
            "type": "string"
          }
        }
      }
    }
  }
}